# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Mapt",
    "description": "Shape file uploader/sharing application",
    "version": "0.1.0"
  },
  "paths": {
    "/layers/": {
      "get": {
        "summary": "Get Layers",
        "operationId": "get_layers_layers__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Layers Layers  Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Layer"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/datasets/shapefile/": {
      "post": {
        "summary": "Upload Shape",
        "description": "Upload a new shape file\n\nThis will eventually be offloaded to the ARQ worker but keep it simple for now and do it synchronously",
        "operationId": "upload_shape_datasets_shapefile__post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_shape_datasets_shapefile__post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dataset"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/datasets/{dataset_id}": {
      "get": {
        "summary": "Get Dataset",
        "operationId": "get_dataset_datasets__dataset_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "integer"
            },
            "name": "dataset_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dataset"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/datasets/": {
      "get": {
        "summary": "Get All Datasets",
        "operationId": "get_all_datasets_datasets__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get All Datasets Datasets  Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Dataset"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Attribute": {
        "title": "Attribute",
        "required": ["name", "display", "type"],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "display": {
            "title": "Display",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        }
      },
      "Body_upload_shape_datasets_shapefile__post": {
        "title": "Body_upload_shape_datasets_shapefile__post",
        "required": ["shape_file"],
        "type": "object",
        "properties": {
          "shape_file": {
            "title": "Shape File",
            "type": "string",
            "format": "binary"
          }
        }
      },
      "Dataset": {
        "title": "Dataset",
        "required": ["label", "description", "schema"],
        "type": "object",
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "schema": {
            "title": "Schema",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attribute"
            }
          }
        }
      },
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "Layer": {
        "title": "Layer",
        "required": ["id", "label", "group"],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "group": {
            "title": "Group",
            "type": "string"
          }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": ["loc", "msg", "type"],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    }
  }
}